third_party/WebKit/LayoutTests/fast/notifications/notifications-cancel-request-permission-expected.txt

CONSOLE ERROR: line 16: Uncaught TypeError: undefined is not a function
Requesting and cancelling notification permission...

Pass if notification is cancelled
